blockmig: add option to skip zero blocks in bulk stage
this patch adds an option to skip zero blocks in the
bulk stage of block migration. this is useful if
all unwritten blocks on the target storage read
as zeros (e.g. thin-provisioned devices).

if all zero blocks where explicitly sent the
target at the destination will unnecessarily
grow to 100% size if there is no support for
re-thinning.

Signed-off-by: Peter Lieven <pl@kamp.de>