block: has_zero_init should only return 1 if the device is really empty
bdrv_has_zero_init is currently only used in qemu-img convert to
check if a destination is zero initialized. so its used as indicator
if the whole destination device reads as zero. consequently zero
blocks are not writting if has_zero_init returns 1.
in the case of qemu-img convert this question
is easy to answer since the the destination device will
be freshly created before it is used.

but there are already exceptions to this. for instance iscsi
targets are only checked if they meet the size requirements on bdrv_create.
the same applies for host devices. for host devices has_zero_init
return 0. but for iscsi in theory there is a flag that indicates
if unallocated blocks read as zero.

a next potential problem could be sparse raw files in the
case the filesystem does not support sparse files. in this case
has_zero_init is 1 altough we do not know for sure that all
blocks read as zeroes.

to avoid problems we additionally check if all blocks of an device
are unallocated. this also enables us to use bdrv_has_zero_init()
outside of the qemu-img convert context.

as soon as the bdrv flags extension is ready, the allocation
check should be changed to check for BDRV_IS_ZERO.

Signed-off-by: Peter Lieven <pl@kamp.de>