block: introduce BDRV_O_SEQUENTIAL
this patch introduces a new flag to indicate that we are going to sequentially
read from a file and do not plan to reread/reuse the data after it has been read.

The current use of this flag is to open the source(s) of a qemu-img convert
process. If a protocol from block/raw-posix.c is used posix_fadvise is utilized
to advise to the kernel that we are going to read sequentially from the
file and a POSIX_FADV_DONTNEED advise is issued after each write to indicate
that there is no advantage keeping the blocks in the buffers. While the
first seems to offer a slight performance benefit the latter option avoids
that older data is swapped out to have the data unnecessarily buffered.

Signed-off-by: Peter Lieven <pl@kamp.de>

Conflicts:

	include/block/block.h

Signed-off-by: Peter Lieven <pl@kamp.de>